dify_plugin/pymupdf4llm/requirements.txt
dify_plugin>=0.4.0,<0.7.0
pymupdf4llm==0.3.4
markdown==3.9
